services/query/requirements.txt
numpy==2.1.1
opencv-python==4.10.0.84
pillow==10.4.0
grpcio==1.66.1
grpcio-tools==1.66.1
protobuf==5.27.2
confluent-kafka==2.5.3
